Index

LagrangianPerturbationTheory.CCLCosmology

LagrangianPerturbationTheory.InitialConditionsWebsky

LagrangianPerturbationTheory.InterpolatedCosmology

LagrangianPerturbationTheory.growth_factor

LagrangianPerturbationTheory.lattice_0

LagrangianPerturbationTheory.scale_factor_of_chi

LagrangianPerturbationTheory.CCLCosmology — Method

CCLCosmology(T; kwargs...)

Returns a wrapped [pyccl cosmology][1] object with output type T and cosmological parameters as keyword arguments. For example, a standard Planck 2018 cosmology can be configured with

cclcosmo = CCLCosmology(Float32; 
    Omega_c=0.2589, Omega_b=0.0486, h=0.6774, sigma8=0.8159,
    n_s=0.9667, transfer_function="boltzmann_camb")

[1]: https://ccl.readthedocs.io/en/latest/api/pyccl.core.html#pyccl.core.Cosmology

LagrangianPerturbationTheory.InitialConditionsWebsky — Method

InitialConditionsWebsky(::Type{LPT}, grid_spacing, cosmo, field)

Returns a wrapper around an array field in Lagrangian coordinates, defined by a kind of Lagrangian perturbation theory LPT, grid spacing, and cosmo. For Websky, this kind of field is typically something like $\delta_0(\vec{q})$, the Lagrangian density at $z=0$.

LagrangianPerturbationTheory.InterpolatedCosmology — Method

InterpolatedCosmology(cosmo)

Given a different cosmology object (at the moment, only CCLCosmology), construct a set of interpolators that provide the quantities needed for LPT.

cclcosmo = CCLCosmology(Float32; 
    Omega_c=0.2589, Omega_b=0.0486, h=0.6774, sigma8=0.8159,
    n_s=0.9667, transfer_function="boltzmann_camb")
cosmo = InterpolatedCosmology(cclcosmo)

LagrangianPerturbationTheory.growth_factor — Function

growth_factor(cosmo, a)

Returns the linear growth factor $D(a)$ as a function of cosmo cosmo and scale factor $a$.

LagrangianPerturbationTheory.lattice_0 — Method

lattice_0(ic, i, j, k, oi, oj, ok)

Returns the lattice value $f_0(\vec{q})$ at $z=0$ for some field $f$, where i, j, and k are the Lagrangian lattice indices so that $\vec{q} = (q_i, q_j, q_k)$, and oi, oj, and ok refer to the octant index (Websky is repeated twice in each dimension with the observer at the center, to form eight subvolumes). Octant indices are either -1 or 0.

Do not access the array directly, use this function instead. Python's array format flips the dimensions of arrays in storage.

LagrangianPerturbationTheory.scale_factor_of_chi — Function

scale_factor_of_chi(cosmo, chi)

Returns the scale factor $a(\chi)$ as a function of comoving radial distance $\chi$ (Mpc).