Index

LagrangianPerturbationTheory.CCLCosmology

LagrangianPerturbationTheory.ICFieldWebsky

LagrangianPerturbationTheory.InterpolatedCosmology

LagrangianPerturbationTheory.dndlogm

LagrangianPerturbationTheory.draw_tracer!

LagrangianPerturbationTheory.growth_factor

LagrangianPerturbationTheory.package_tracer_positions

LagrangianPerturbationTheory.scale_factor_of_chi

LagrangianPerturbationTheory.CCLCosmology — Method

CCLCosmology(T; kwargs...)

Returns a wrapped [pyccl cosmology][1] object with output type T and cosmological parameters as keyword arguments. For example, a standard Planck 2018 cosmology can be configured with

cclcosmo = CCLCosmology(Float32; 
    Omega_c=0.2589, Omega_b=0.0486, h=0.6774, sigma8=0.8159,
    n_s=0.9667, transfer_function="boltzmann_camb")

[1]: https://ccl.readthedocs.io/en/latest/api/pyccl.core.html#pyccl.core.Cosmology

LagrangianPerturbationTheory.ICFieldWebsky — Method

ICFieldWebsky

Returns a wrapper around an array field in Lagrangian coordinates, defined by a kind of Lagrangian perturbation theory LPT, grid spacing, and cosmo. For Websky, this kind of field is typically something like $\delta_0(\vec{q})$, the Lagrangian density at $z=0$.

LagrangianPerturbationTheory.InterpolatedCosmology — Method

InterpolatedCosmology(cosmo)

Given a different cosmology object (at the moment, only CCLCosmology), construct a set of interpolators that provide the quantities needed for LPT.

cclcosmo = CCLCosmology(Float32; 
    Omega_c=0.2589, Omega_b=0.0486, h=0.6774, sigma8=0.8159,
    n_s=0.9667, transfer_function="boltzmann_camb")
cosmo = InterpolatedCosmology(cclcosmo)

LagrangianPerturbationTheory.dndlogm — Method

as in pyccl, returns number density per Mpc³ per log₁₀(Mₕ/Msun)

LagrangianPerturbationTheory.draw_tracer! — Method

Octants are indexed -1 and 0.

LagrangianPerturbationTheory.growth_factor — Function

growth_factor(cosmo, a)

Returns the linear growth factor $D(a)$ as a function of cosmo cosmo and scale factor $a$.

LagrangianPerturbationTheory.package_tracer_positions — Method

package_tracer_positions(vector_of_vectors_of_svector_positions)

Convert a triply-nested linearly-indexed of SVector{3,Quantity{T}} (unitful positions) into an Array{T,2} of (3, total_elements) to be stored to disk, without units. Lengths are converted to Mpc.

LagrangianPerturbationTheory.scale_factor_of_chi — Function

scale_factor_of_chi(cosmo, chi)

Returns the scale factor $a(\chi)$ as a function of comoving radial distance $\chi$ (Mpc).